derive_more

Module derive

Module containing derive definitions only, without their corresponding traits.

Use it in your import paths, if you don’t want to import traits, but only macros.

Derive Macros

Add

What #[derive(Add)] generates

AddAssign

What #[derive(AddAssign)] generates

AsMut

What #[derive(AsRef)] generates

AsRef

What #[derive(AsRef)] generates

Binary

What #[derive(Display)] generates

BitAnd

What #[derive(Add)] generates

BitAndAssign

What #[derive(AddAssign)] generates

BitOr

What #[derive(Add)] generates

BitOrAssign

What #[derive(AddAssign)] generates

BitXor

What #[derive(Add)] generates

BitXorAssign

What #[derive(AddAssign)] generates

Constructor

What #[derive(Constructor)] generates

Debug

What #[derive(Debug)] generates

Deref

Using #[derive(Deref)]

DerefMut

What #[derive(DerefMut)] generates

Display

What #[derive(Display)] generates

Div

What #[derive(Mul)] generates

DivAssign

What #[derive(MulAssign)] generates

Error

Using #[derive(Error)]

From

What #[derive(From)] generates

FromStr

What #[derive(FromStr)] generates

Index

What #[derive(Index)] generates

IndexMut

What #[derive(IndexMut)] generates

Into

What #[derive(Into)] generates

IntoIterator

Using #[derive(IntoIterator)]

IsVariant

What #[derive(IsVariant)] generates

LowerExp

What #[derive(Display)] generates

LowerHex

What #[derive(Display)] generates

Mul

What #[derive(Mul)] generates

MulAssign

What #[derive(MulAssign)] generates

Neg

What #[derive(Not)] generates

Not

What #[derive(Not)] generates

Octal

What #[derive(Display)] generates

Pointer

What #[derive(Display)] generates

Product

Using #[derive(Sum)]

Rem

What #[derive(Mul)] generates

RemAssign

What #[derive(MulAssign)] generates

Shl

What #[derive(Mul)] generates

ShlAssign

What #[derive(MulAssign)] generates

Shr

What #[derive(Mul)] generates

ShrAssign

What #[derive(MulAssign)] generates

Sub

What #[derive(Add)] generates

SubAssign

What #[derive(AddAssign)] generates

Sum

Using #[derive(Sum)]

TryFrom

What #[derive(TryFrom)] generates

TryInto

What #[derive(TryInto)] generates

TryUnwrap

What #[derive(TryUnwrap)] generates

Unwrap

What #[derive(Unwrap)] generates

UpperExp

What #[derive(Display)] generates

UpperHex

What #[derive(Display)] generates