derive_more
Module with_trait

Module containing derive definitions with their corresponding traits along.

Use it in your import paths, if you do want to import derives along with their traits.

Derive Macros

Add

What #[derive(Add)] generates

AddAssign

What #[derive(AddAssign)] generates

AsMut

What #[derive(AsRef)] generates

AsRef

What #[derive(AsRef)] generates

Binary

What #[derive(Display)] generates

BitAnd

What #[derive(Add)] generates

BitAndAssign

What #[derive(AddAssign)] generates

BitOr

What #[derive(Add)] generates

BitOrAssign

What #[derive(AddAssign)] generates

BitXor

What #[derive(Add)] generates

BitXorAssign

What #[derive(AddAssign)] generates

Deref

Using #[derive(Deref)]

DerefMut

What #[derive(DerefMut)] generates

Display

What #[derive(Display)] generates

From

What #[derive(From)] generates

FromStr

What #[derive(FromStr)] generates

Index

What #[derive(Index)] generates

IndexMut

What #[derive(IndexMut)] generates

Into

What #[derive(Into)] generates

IntoIterator

Using #[derive(IntoIterator)]

LowerExp

What #[derive(Display)] generates

LowerHex

What #[derive(Display)] generates

Neg

What #[derive(Not)] generates

Not

What #[derive(Not)] generates

Octal

What #[derive(Display)] generates

Pointer

What #[derive(Display)] generates

Sub

What #[derive(Add)] generates

SubAssign

What #[derive(AddAssign)] generates

UpperExp

What #[derive(Display)] generates

UpperHex

What #[derive(Display)] generates