Crate derive_more

derive_more

Rust has lots of builtin traits that are implemented for its basic types, such as Add, Not or From. However, when wrapping these types inside your own structs or enums you lose the implementations of these traits and are required to recreate them. This is especially annoying when your own structures are very simple, such as when using the commonly advised newtype pattern (e.g. MyInt(i32)).

This library tries to remove these annoyances and the corresponding boilerplate code. It does this by allowing you to derive lots of commonly used traits for both structs and enums.

Example code

By using this library the following code just works:

#[macro_use]
extern crate derive_more;

#[derive(Debug, Eq, PartialEq, From, Add)]
struct MyInt(i32);

#[derive(Debug, Eq, PartialEq, From, Into, Constructor, Mul)]
struct Point2D {
    x: i32,
    y: i32,
}

#[derive(Debug, Eq, PartialEq, From, Add)]
enum MyEnum {
    Int(i32),
    UnsignedInt(u32),
    Nothing,
}

fn main() {
    let my_11 = MyInt(5) + 6.into();
    assert_eq!(MyInt(11), MyInt(5) + 6.into());
    assert_eq!(Point2D { x: 5, y: 6 } * 10, (50, 60).into());
    assert_eq!((5, 6), Point2D { x: 5, y: 6 }.into());
    assert_eq!(Point2D { x: 5, y: 6 }, Point2D::new(5, 6));
    assert_eq!(MyEnum::Int(15), (MyEnum::Int(8) + 7.into()).unwrap())
}

The derivable traits

Below are all the traits that you can derive using this library. Some trait derivations are so similar that the further documentation will only show a single one of them. You can recognize these by the "-like" suffix in their name. The trait name before that will be the only one that is used throughout the further documentation.

NOTE: You still have to derive each trait separately. So #[derive(Mul)] doesn't automatically derive Div as well. To derive both you should do #[derive(Mul, Div)]

Conversion traits

These are traits that are used to convert automatically between types.

From

Into

FromStr

TryInto

Formatting traits

These traits are used for converting a struct to a string in different ways.

Display-like, contains Display, Binary, Octal, LowerHex, UpperHex, LowerExp, UpperExp, Pointer

Operators

These are traits that can be used for operator overloading.

Index

Deref

Not-like, contains Not and Neg

Add-like, contains Add, Sub, BitAnd, BitOr and BitXor

Mul-like, contains Mul, Div, Rem, Shr and Shl

IndexMut

DerefMut

AddAssign-like, contains AddAssign, SubAssign, BitAndAssign, BitOrAssign and BitXorAssign

MulAssign-like, contains MulAssign, DivAssign, RemAssign, ShrAssign and ShlAssign

Static methods

These don't derive traits, but derive static methods instead.

Constructor, this derives a new method that can be used as a constructor. This is very basic if you need more customization for your constructor, check out the derive-new crate.

Generated code

It is important to understand what code gets generated when using one of the derives from this crate. That is why the links below explain what code gets generated for a trait for each group from before.

#[derive(From)]

#[derive(Into)]

#[derive(FromStr)]

#[derive(TryInto)]

#[derive(Display)]

#[derive(Index)]

#[derive(Deref)]

#[derive(Not)]

#[derive(Add)]

#[derive(Mul)]

#[derive(IndexMut)]

#[derive(DerefMut)]

#[derive(AddAssign)]

#[derive(MulAssign)]

#[derive(Constructor)]

If you want to be sure what code is generated for your specific type I recommend using the cargo-expand utility. This will show you your code with all macros and derives expanded.

Installation

This library requires Rust 1.15 or higher, so this needs to be installed. Then add the following to Cargo.toml:

[dependencies]
derive_more = "0.13.0"

And this to the top of your Rust file:

#[macro_use]
extern crate derive_more;

This crate support no_std through the no_std feature. So use the following instead if you want to use it in a no_std environment.

# Example Cargo.toml
[dependencies]
derive_more = {version = "0.13.0", default-features = false, features=["no_std"]}