Crate derive_more

Modules

derive

Module containing macro definitions only, without corresponding traits.

Structs

UnitError

Error returned by the derived implementations when an arithmetic or logic operation is invoked on a unit-like variant of an enum.

WrongVariantError

Error returned by the derived implementations when an arithmetic or logic operation is invoked on mismatched enum variants.

Enums

BinaryError

Possible errors returned by the derived implementations of binary arithmetic or logic operations.

Derive Macros

Add

What #[derive(Add)] generates

AddAssign

What #[derive(AddAssign)] generates

Binary

What #[derive(Display)] generates

BitAnd

What #[derive(Add)] generates

BitAndAssign

What #[derive(AddAssign)] generates

BitOr

What #[derive(Add)] generates

BitOrAssign

What #[derive(AddAssign)] generates

BitXor

What #[derive(Add)] generates

BitXorAssign

What #[derive(AddAssign)] generates

Display

What #[derive(Display)] generates

Div

What #[derive(Mul)] generates

DivAssign

What #[derive(MulAssign)] generates

LowerExp

What #[derive(Display)] generates

LowerHex

What #[derive(Display)] generates

Mul

What #[derive(Mul)] generates

MulAssign

What #[derive(MulAssign)] generates

Neg

What #[derive(Not)] generates

Not

What #[derive(Not)] generates

Octal

What #[derive(Display)] generates

Pointer

What #[derive(Display)] generates

Rem

What #[derive(Mul)] generates

RemAssign

What #[derive(MulAssign)] generates

Shl

What #[derive(Mul)] generates

ShlAssign

What #[derive(MulAssign)] generates

Shr

What #[derive(Mul)] generates

ShrAssign

What #[derive(MulAssign)] generates

Sub

What #[derive(Add)] generates

SubAssign

What #[derive(AddAssign)] generates

UpperExp

What #[derive(Display)] generates

UpperHex

What #[derive(Display)] generates